[Feat] #32 - fix CGFlaot literals

+++ Puzzle-iOS/Puzzle-iOS/Resource/UIComponents/PuzzleDropdownTableView.swift
@@ -89,7 +89,7 @@ extension PuzzleDropdownTableView: UITableViewDelegate {
     }
     
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return CGFloat(NumberLiterals.DropDown.dropDownHeight)
+        return NumberLiterals.DropDown.dropDownHeight
     }
 }
